v1.3
1. New push system, AI will be scattered stations when push, and will choose a more appropriate time to push the tower. 2. Improved warding system.3. Supported the captain mode.

+++ ability_item_usage_winter_wyvern.lua
@@ -475,6 +475,11 @@ Consider[4]=function()
 		end
 	end
 	
+	if(enemys~=nil and #enemys==1)
+	then
+		return 0,0
+	end
+	
 	--Try to kill enemy hero
 	if(npcBot:GetActiveMode() ~= BOT_MODE_RETREAT ) 
 	then
